Rename ServerErrorMessageBuilder (#20647)
